change the c_frontend APIs with corresponding C++ frontend APIs. 1. split the tdi_table.h into two files.    tdi_table.h with table_hdl and tdi_table_info.h with table_info_hdl. 2. similar split APIs into two files.    tdi_table_c.h with table_hdl and tdi_table_info_c.cpp with table_info_hdl.
